Revert "int ver"
This reverts commit ca2e3c145bd9f7a28e08b3bb947a0944909a732e.

+++ backend/pkg/scheduler/scheduler.go
@@ -39,7 +39,7 @@ type Scheduler struct {
 type Params struct {
 	Namespaces                    []string
 	CisDockerBenchmarkScanEnabled bool
-	MaxScanParallelism            int
+	MaxScanParallelism            int64
 	Interval                      time.Duration
 	StartTime                     time.Time
 	SingleScan                    bool
@@ -86,7 +86,7 @@ func (s *Scheduler) Init() {
 	s.Schedule(&Params{
 		Namespaces:                    scanConfig.Namespaces,
 		CisDockerBenchmarkScanEnabled: scanConfig.CisDockerBenchmarkScanEnabled,
-		MaxScanParallelism:            int(scanConfig.MaxScanParallelism),
+		MaxScanParallelism:            scanConfig.MaxScanParallelism,
 		Interval:                      time.Duration(sched.Interval),
 		StartTime:                     startTime,
 		SingleScan:                    scanConfig.ScanConfigType().ScheduleScanConfigType() == SingleScheduleScanConfig,
